VideoScreen moved to corridor
When VideoScreen is added I can not get the Player to walk into the cell. Do not know why not.  Moved VideoScreen into corridor so can work on getting Video to Play on Trigger and not just start on awake/play scene.  Have made box collider big so easy to collide with.  Have tagged FPC as Player.