Copy of Door made to try to get it to close
Copy made of cell door 004. Made it a child of an empty game object called DoorController.  Attached a script called DoorController to the Empty Game Object called DoorController.  Script does move the door on pressing 'E' at moment.  But the arc it moves in is not correct and needs to be fixed.  Also need to fix the trigger for the door so not pressing 'E' but to something else better for the user experience.